[addons] rename addon dialogs to use namespace kodi::gui::dialogs::..
Change to make the different gui parts more clear.
