+++ xbmc/addons/interfaces/GUI/CMakeLists.txt
@@ -1,14 +1,4 @@
 set(SOURCES General.cpp
-            DialogContextMenu.cpp
-            DialogExtendedProgressBar.cpp
-            DialogFileBrowser.cpp
-            DialogKeyboard.cpp
-            DialogNumeric.cpp
-            DialogOK.cpp
-            DialogProgress.cpp
-            DialogSelect.cpp
-            DialogTextViewer.cpp
-            DialogYesNo.cpp
             ListItem.cpp
             Window.cpp
             AddonCallbacksGUI.cpp
@@ -16,16 +6,6 @@ set(SOURCES General.cpp
             AddonGUIWindow.cpp)
 
 set(HEADERS General.h
-            DialogContextMenu.h
-            DialogExtendedProgressBar.h
-            DialogFileBrowser.h
-            DialogKeyboard.h
-            DialogNumeric.h
-            DialogOK.h
-            DialogProgress.h
-            DialogSelect.h
-            DialogTextViewer.h
-            DialogYesNo.h
             ListItem.h
             Window.h
             AddonCallbacksGUI.h
