+++ xbmc/addons/interfaces/GUI/General.cpp
@@ -20,16 +20,16 @@
 
 #include "General.h"
 #include "controls/Button.h"
-#include "DialogContextMenu.h"
-#include "DialogExtendedProgressBar.h"
-#include "DialogFileBrowser.h"
-#include "DialogKeyboard.h"
-#include "DialogNumeric.h"
-#include "DialogOK.h"
-#include "DialogProgress.h"
-#include "DialogSelect.h"
-#include "DialogTextViewer.h"
-#include "DialogYesNo.h"
+#include "dialogs/ContextMenu.h"
+#include "dialogs/ExtendedProgressBar.h"
+#include "dialogs/FileBrowser.h"
+#include "dialogs/Keyboard.h"
+#include "dialogs/Numeric.h"
+#include "dialogs/OK.h"
+#include "dialogs/Progress.h"
+#include "dialogs/Select.h"
+#include "dialogs/TextViewer.h"
+#include "dialogs/YesNo.h"
 #include "ListItem.h"
 #include "Window.h"
 #include "addons/kodi-addon-dev-kit/include/kodi/gui/General.h"
