+++ xbmc/addons/interfaces/GUI/dialogs/CMakeLists.txt
@@ -1,0 +1,23 @@
+set(SOURCES ContextMenu.cpp
+            ExtendedProgressBar.cpp
+            FileBrowser.cpp
+            Keyboard.cpp
+            Numeric.cpp
+            OK.cpp
+            Progress.cpp
+            Select.cpp
+            TextViewer.cpp
+            YesNo.cpp)
+
+set(HEADERS ContextMenu.h
+            ExtendedProgressBar.h
+            FileBrowser.h
+            Keyboard.h
+            Numeric.h
+            OK.h
+            Progress.h
+            Select.h
+            TextViewer.h
+            YesNo.h)
+
+core_add_library(addons_interface_gui_dialogs)
