+++ xbmc/addons/interfaces/GUI/dialogs/FileBrowser.h
@@ -40,7 +40,7 @@ namespace ADDON
    * To hold functions not related to a instance type and usable for
    * every add-on type.
    *
-   * Related add-on header is "./xbmc/addons/kodi-addon-dev-kit/include/kodi/gui/DialogFileBrowser.h"
+   * Related add-on header is "./xbmc/addons/kodi-addon-dev-kit/include/kodi/gui/dialogs/FileBrowser.h"
    */
   struct Interface_GUIDialogFileBrowser
   {
