+++ xbmc/addons/interfaces/GUI/dialogs/Numeric.h
@@ -19,6 +19,8 @@
  *
  */
 
+#include <time.h>
+
 extern "C"
 {
 
@@ -33,9 +35,9 @@ namespace ADDON
    * To hold functions not related to a instance type and usable for
    * every add-on type.
    *
-   * Related add-on header is "./xbmc/addons/kodi-addon-dev-kit/include/kodi/gui/DialogOK.h"
+   * Related add-on header is "./xbmc/addons/kodi-addon-dev-kit/include/kodi/gui/dialogs/Numeric.h"
    */
-  struct Interface_GUIDialogOK
+  struct Interface_GUIDialogNumeric
   {
     static void Init(AddonGlobalInterface* addonInterface);
     static void DeInit(AddonGlobalInterface* addonInterface);
@@ -43,16 +45,22 @@ namespace ADDON
     /*!
      * @brief callback functions from add-on to kodi
      *
-     * @note For add of new functions use the "_" style to identify direct a
-     * add-on callback function. Everything with CamelCase is only for the
-     * usage in Kodi only.
+     * @note To add a new function use the "_" style to directly identify an
+     * add-on callback function. Everything with CamelCase is only to be used
+     * in Kodi.
      *
      * The parameter `kodiBase` is used to become the pointer for a `CAddonDll`
      * class.
      */
     //@{
-    static void show_and_get_input_single_text(void* kodiBase, const char *heading, const char *text);
-    static void show_and_get_input_line_text(void* kodiBase, const char *heading, const char *line0, const char *line1, const char *line2);
+    static bool show_and_verify_new_password(void* kodiBase, char** password);
+    static int show_and_verify_password(void* kodiBase, const char* password, const char *heading, int retries);
+    static bool show_and_verify_input(void* kodiBase, const char* verify_in, char** verify_out, const char* heading, bool verify_input);
+    static bool show_and_get_time(void* kodiBase, tm *time, const char *heading);
+    static bool show_and_get_date(void* kodiBase, tm *date, const char *heading);
+    static bool show_and_get_ip_address(void* kodiBase, const char* ip_address_in, char** ip_address_out, const char *heading);
+    static bool show_and_get_number(void* kodiBase, const char* number_in, char** number_out, const char *heading, unsigned int auto_close_ms);
+    static bool show_and_get_seconds(void* kodiBase, const char* time_in, char** time_out, const char *heading);
     //@}
   };
