+++ xbmc/addons/interfaces/GUI/dialogs/Progress.h
@@ -33,9 +33,9 @@ namespace ADDON
    * To hold functions not related to a instance type and usable for
    * every add-on type.
    *
-   * Related add-on header is "./xbmc/addons/kodi-addon-dev-kit/include/kodi/gui/DialogKeyboard.h"
+   * Related add-on header is "./xbmc/addons/kodi-addon-dev-kit/include/kodi/gui/dialogs/Progress.h"
    */
-  struct Interface_GUIDialogKeyboard
+  struct Interface_GUIDialogProgress
   {
     static void Init(AddonGlobalInterface* addonInterface);
     static void DeInit(AddonGlobalInterface* addonInterface);
@@ -51,16 +51,19 @@ namespace ADDON
      * class.
      */
     //@{
-    static bool show_and_get_input_with_head(void* kodiBase, const char* text_in, char** text_out, const char* heading, bool allow_empty_result, bool hidden_input, unsigned int auto_close_ms);
-    static bool show_and_get_input(void* kodiBase, const char* text_in, char** text_out, bool allow_empty_result, unsigned int auto_close_ms);
-    static bool show_and_get_new_password_with_head(void* kodiBase, const char* password_in, char** password_out, const char* heading, bool allow_empty_result, unsigned int auto_close_ms);
-    static bool show_and_get_new_password(void* kodiBase, const char* password_in, char** password_out, unsigned int auto_close_ms);
-    static bool show_and_verify_new_password_with_head(void* kodiBase, char** password_out, const char* heading, bool allowEmpty, unsigned int auto_close_ms);
-    static bool show_and_verify_new_password(void* kodiBase, char** password_out, unsigned int auto_close_ms);
-    static int show_and_verify_password(void* kodiBase, const char* password_in, char** password_out, const char* heading, int retries, unsigned int auto_close_ms);
-    static bool show_and_get_filter(void* kodiBase, const char* text_in, char** text_out, bool searching, unsigned int auto_close_ms);
-    static bool send_text_to_active_keyboard(void* kodiBase, const char* text, bool close_keyboard);
-    static bool is_keyboard_activated(void* kodiBase);
+    static void* new_dialog(void* kodiBase);
+    static void delete_dialog(void* kodiBase, void* handle);
+    static void open(void* kodiBase, void* handle);
+    static void set_heading(void* kodiBase, void* handle, const char* heading);
+    static void set_line(void* kodiBase, void* handle, unsigned int line, const char* text);
+    static void set_can_cancel(void* kodiBase, void* handle, bool canCancel);
+    static bool is_canceled(void* kodiBase, void* handle);
+    static void set_percentage(void* kodiBase, void* handle, int percentage);
+    static int get_percentage(void* kodiBase, void* handle);
+    static void show_progress_bar(void* kodiBase, void* handle, bool bOnOff);
+    static void set_progress_max(void* kodiBase, void* handle, int max);
+    static void set_progress_advance(void* kodiBase, void* handle, int nSteps);
+    static bool abort(void* kodiBase, void* handle);
     //@}
   };
