+++ xbmc/addons/kodi-addon-dev-kit/include/kodi/gui/dialogs/ContextMenu.h
@@ -19,33 +19,35 @@
  *
  */
 
-#include "definitions.h"
-#include "../AddonBase.h"
+#include "../definitions.h"
+#include "../../AddonBase.h"
 
 namespace kodi
 {
 namespace gui
 {
+namespace dialogs
+{
 
   //============================================================================
   ///
-  /// \defgroup cpp_kodi_gui_DialogContextMenu Dialog Context Menu
+  /// \defgroup cpp_kodi_gui_dialogs_ContextMenu Dialog Context Menu
   /// \ingroup cpp_kodi_gui
-  /// @brief \cpp_namespace{ kodi::gui::DialogContextMenu }
+  /// @brief \cpp_namespace{ kodi::gui::dialogs::ContextMenu }
   /// **Context menu dialog**
   ///
   /// The function listed below permits the call of a dialogue as context menu to
   /// select of an entry as a key
   ///
-  /// It has the header \ref DialogContextMenu.h "#include <kodi/gui/DialogContextMenu.h>"
+  /// It has the header \ref ContextMenu.h "#include <kodi/gui/dialogs/ContextMenu.h>"
   /// be included to enjoy it.
   ///
   ///
-  namespace DialogContextMenu
+  namespace ContextMenu
   {
     //==========================================================================
     ///
-    /// \ingroup cpp_kodi_gui_DialogContextMenu
+    /// \ingroup cpp_kodi_gui_dialogs_ContextMenu
     /// @brief Show a context menu dialog about given parts.
     ///
     /// @param[in] heading   Dialog heading name
@@ -57,7 +59,7 @@ namespace gui
     ///
     /// **Example:**
     /// ~~~~~~~~~~~~~{.cpp}
-    /// #include <kodi/gui/DialogContextMenu.h>
+    /// #include <kodi/gui/dialogs/ContextMenu.h>
     ///
     /// const std::vector<std::string> entries
     /// {
@@ -68,7 +70,7 @@ namespace gui
     ///   "Test 5"
     /// };
     ///
-    /// int selected = kodi::gui::DialogContextMenu::Show("Test selection", entries);
+    /// int selected = kodi::gui::dialogs::ContextMenu::Show("Test selection", entries);
     /// if (selected < 0)
     ///   fprintf(stderr, "Item selection canceled\n");
     /// else
@@ -91,5 +93,6 @@ namespace gui
     //--------------------------------------------------------------------------
   };
 
+} /* namespace dialogs */
 } /* namespace gui */
 } /* namespace kodi */
