+++ xbmc/addons/kodi-addon-dev-kit/include/kodi/gui/dialogs/ExtendedProgress.h
@@ -19,22 +19,24 @@
  *
  */
 
-#include "definitions.h"
-#include "../AddonBase.h"
+#include "../definitions.h"
+#include "../../AddonBase.h"
 
 namespace kodi
 {
 namespace gui
 {
+namespace dialogs
+{
 
   //============================================================================
   ///
-  /// \defgroup cpp_kodi_gui_CDialogExtendedProgress Dialog Extended Progress
+  /// \defgroup cpp_kodi_gui_dialogs_CExtendedProgress Dialog Extended Progress
   /// \ingroup cpp_kodi_gui
-  /// @brief \cpp_class{ kodi::gui::CDialogExtendedProgress }
+  /// @brief \cpp_class{ kodi::gui::dialogs::ExtendedProgress }
   /// **Progress dialog shown for background work**
   ///
-  /// The with \ref DialogExtendedProgress.h "#include <kodi/gui/DialogExtendedProgress.h>"
+  /// The with \ref ExtendedProgress.h "#include <kodi/gui/dialogs/ExtendedProgress.h>"
   /// given class are basically used to create Kodi's extended progress.
   ///
   ///
@@ -42,9 +44,9 @@ namespace gui
   ///
   /// **Example:**
   /// ~~~~~~~~~~~~~{.cpp}
-  /// #include <kodi/gui/DialogExtendedProgress.h>
+  /// #include <kodi/gui/dialogs/ExtendedProgress.h>
   ///
-  /// kodi::gui::CDialogExtendedProgress *ext_progress = new kodi::gui::CDialogExtendedProgress("Test Extended progress");
+  /// kodi::gui::dialogs::CExtendedProgress *ext_progress = new kodi::gui::dialogs::CExtendedProgress("Test Extended progress");
   /// ext_progress->SetText("Test progress");
   /// for (unsigned int i = 0; i < 50; i += 10)
   /// {
@@ -63,17 +65,17 @@ namespace gui
   /// delete ext_progress;
   /// ~~~~~~~~~~~~~
   ///
-  class CDialogExtendedProgress
+  class CExtendedProgress
   {
   public:
     //==========================================================================
     ///
-    /// \ingroup cpp_kodi_gui_CDialogExtendedProgress
+    /// \ingroup cpp_kodi_gui_dialogs_CExtendedProgress
     /// Construct a new dialog
     ///
     /// @param[in] title  Title string
     ///
-    CDialogExtendedProgress(const std::string& title = "")
+    CExtendedProgress(const std::string& title = "")
     {
       using namespace ::kodi::addon;
       m_DialogHandle = CAddonBase::m_interface->toKodi->kodi_gui->dialogExtendedProgress->new_dialog(CAddonBase::m_interface->toKodi->kodiBase, title.c_str());
@@ -84,10 +86,10 @@ namespace gui
 
     //==========================================================================
     ///
-    /// \ingroup cpp_kodi_gui_CDialogExtendedProgress
+    /// \ingroup cpp_kodi_gui_dialogs_CExtendedProgress
     /// Destructor
     ///
-    ~CDialogExtendedProgress()
+    ~CExtendedProgress()
     {
       using namespace ::kodi::addon;
       if (m_DialogHandle)
@@ -97,7 +99,7 @@ namespace gui
 
     //==========================================================================
     ///
-    /// \ingroup cpp_kodi_gui_CDialogExtendedProgress
+    /// \ingroup cpp_kodi_gui_dialogs_CExtendedProgress
     /// @brief Get the used title
     ///
     /// @return Title string
@@ -119,7 +121,7 @@ namespace gui
 
     //==========================================================================
     ///
-    /// \ingroup cpp_kodi_gui_CDialogExtendedProgress
+    /// \ingroup cpp_kodi_gui_dialogs_CExtendedProgress
     /// @brief To set the title of dialog
     ///
     /// @param[in] title     Title string
@@ -133,7 +135,7 @@ namespace gui
 
     //==========================================================================
     ///
-    /// \ingroup cpp_kodi_gui_CDialogExtendedProgress
+    /// \ingroup cpp_kodi_gui_dialogs_CExtendedProgress
     /// @brief Get the used text information string
     ///
     /// @return Text string
@@ -155,7 +157,7 @@ namespace gui
 
     //==========================================================================
     ///
-    /// \ingroup cpp_kodi_gui_CDialogExtendedProgress
+    /// \ingroup cpp_kodi_gui_dialogs_CExtendedProgress
     /// @brief To set the used text information string
     ///
     /// @param[in] text         information text to set
@@ -169,7 +171,7 @@ namespace gui
 
     //==========================================================================
     ///
-    /// \ingroup cpp_kodi_gui_CDialogExtendedProgress
+    /// \ingroup cpp_kodi_gui_dialogs_CExtendedProgress
     /// @brief To ask dialog is finished
     ///
     /// @return True if on end
@@ -183,7 +185,7 @@ namespace gui
 
     //==========================================================================
     ///
-    /// \ingroup cpp_kodi_gui_CDialogExtendedProgress
+    /// \ingroup cpp_kodi_gui_dialogs_CExtendedProgress
     /// @brief Mark progress finished
     ///
     void MarkFinished()
@@ -195,7 +197,7 @@ namespace gui
 
     //==========================================================================
     ///
-    /// \ingroup cpp_kodi_gui_CDialogExtendedProgress
+    /// \ingroup cpp_kodi_gui_dialogs_CExtendedProgress
     /// @brief Get the current progress position as percent
     ///
     /// @return Position
@@ -209,7 +211,7 @@ namespace gui
 
     //==========================================================================
     ///
-    /// \ingroup cpp_kodi_gui_CDialogExtendedProgress
+    /// \ingroup cpp_kodi_gui_dialogs_CExtendedProgress
     /// @brief To set the current progress position as percent
     ///
     /// @param[in] percentage   Position to use from 0.0 to 100.0
@@ -223,7 +225,7 @@ namespace gui
 
     //==========================================================================
     ///
-    /// \ingroup cpp_kodi_gui_CDialogExtendedProgress
+    /// \ingroup cpp_kodi_gui_dialogs_CExtendedProgress
     /// @brief To set progress position with predefined places
     ///
     /// @param[in] currentItem    Place position to use
@@ -240,5 +242,6 @@ namespace gui
     void* m_DialogHandle;
   };
 
+} /* namespace dialogs */
 } /* namespace gui */
 } /* namespace kodi */
